+++ tests/testthat/test-bounds.r
@@ -2,13 +2,13 @@ context("Bounds")
 
 test_that("rescale_mid returns correct results", {
   x <- c(-1, 0, 1)
-  
+
   expect_equal(rescale_mid(x), c(0, 0.5, 1))
   expect_equal(rescale_mid(x, mid = -1), c(0.5, 0.75, 1))
   expect_equal(rescale_mid(x, mid = 1), c(0, 0.25, 0.5))
-  
+
   expect_equal(rescale_mid(x, mid = 1, to = c(0, 10)), c(0, 2.5, 5))
-  expect_equal(rescale_mid(x, mid = 1, to = c(8, 10)), c(8, 8.5, 9))  
+  expect_equal(rescale_mid(x, mid = 1, to = c(8, 10)), c(8, 8.5, 9))
 })
 
 
@@ -23,13 +23,27 @@ test_that("resacle_max returns correct results", {
 
 test_that("zero range inputs return mid range", {
   expect_that(rescale(0), equals(0.5))
-  expect_that(rescale(c(0, 0)), equals(c(0.5, 0.5)))  
+  expect_that(rescale(c(0, 0)), equals(c(0.5, 0.5)))
 })
 
 
 test_that("censor and squish ignore infinite values", {
   expect_equal(squish(c(1, Inf)), c(1, Inf))
   expect_equal(censor(c(1, Inf)), c(1, Inf))
-  
-  
+
+
+})
+
+test_that("scaling is possible with dates and times", {
+  dates <- as.Date(c("2010-01-01", "2010-01-03", "2010-01-05", "2010-01-07"))
+  expect_equal(rescale(dates, from = c(dates[1], dates[4])), seq(0,1,1/3))
+  expect_equal(rescale_mid(dates, mid = dates[3])[3], 0.5)
+
+  dates <- as.POSIXct(c("2010-01-01 01:40:40",
+                        "2010-01-01 03:40:40",
+                        "2010-01-01 05:40:40",
+                        "2010-01-01 07:40:40"))
+  expect_equal(rescale(dates, from = c(dates[1], dates[4])), seq(0, 1, 1/3))
+  expect_equal(rescale_mid(dates, mid = dates[3])[3], 0.5)
+
 })